refactor: move all domain/lemmy use cases to separate module
