+++ domain/lemmy/usecase/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+    id("com.livefast.eattrash.kotlinMultiplatform")
+    id("com.livefast.eattrash.androidTest")
+    alias(libs.plugins.kotlinx.kover)
+}
+
+kotlin {
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                implementation(libs.kodein)
+
+                implementation(projects.domain.lemmy.data)
+                implementation(projects.domain.lemmy.repository)
+            }
+        }
+    }
+}
