+++ domain/lemmy/usecase/src/commonMain/kotlin/com/livefast/eattrash/raccoonforlemmy/domain/lemmy/usecase/DefaultGetSiteSupportsMediaListUseCase.kt
@@ -1,8 +1,8 @@
-package com.livefast.eattrash.raccoonforlemmy.domain.lemmy.repository
+package com.livefast.eattrash.raccoonforlemmy.domain.lemmy.usecase
 
-internal class DefaultGetSiteSupportsHiddenPostsUseCase(
+internal class DefaultGetSiteSupportsMediaListUseCase(
     private val isSiteVersionAtLeastUseCase: IsSiteVersionAtLeastUseCase,
-) : GetSiteSupportsHiddenPostsUseCase {
+) : GetSiteSupportsMediaListUseCase {
     companion object {
         const val THRESHOLD_MAJOR = 0
         const val THRESHOLD_MINOR = 19
