+++ domain/lemmy/usecase/src/commonMain/kotlin/com/livefast/eattrash/raccoonforlemmy/domain/lemmy/usecase/GetSiteSupportsHiddenPostsUseCase.kt
@@ -1,4 +1,4 @@
-package com.livefast.eattrash.raccoonforlemmy.domain.lemmy.repository
+package com.livefast.eattrash.raccoonforlemmy.domain.lemmy.usecase
 
 interface GetSiteSupportsHiddenPostsUseCase {
     suspend operator fun invoke(): Boolean
