+++ domain/lemmy/usecase/src/commonMain/kotlin/com/livefast/eattrash/raccoonforlemmy/domain/lemmy/usecase/GetSiteSupportsMediaListUseCase.kt
@@ -1,4 +1,4 @@
-package com.livefast.eattrash.raccoonforlemmy.domain.lemmy.repository
+package com.livefast.eattrash.raccoonforlemmy.domain.lemmy.usecase
 
 interface GetSiteSupportsMediaListUseCase {
     suspend operator fun invoke(): Boolean
