+++ domain/lemmy/usecase/src/commonMain/kotlin/com/livefast/eattrash/raccoonforlemmy/domain/lemmy/usecase/GetSortTypesUseCase.kt
@@ -1,4 +1,4 @@
-package com.livefast.eattrash.raccoonforlemmy.domain.lemmy.repository
+package com.livefast.eattrash.raccoonforlemmy.domain.lemmy.usecase
 
 import com.livefast.eattrash.raccoonforlemmy.domain.lemmy.data.SortType
