+++ feature/settings/build.gradle.kts
@@ -30,6 +30,7 @@ kotlin {
 
                 implementation(projects.domain.lemmy.data)
                 implementation(projects.domain.lemmy.repository)
+                implementation(projects.domain.lemmy.usecase)
                 implementation(projects.domain.identity)
 
                 implementation(projects.unit.web)
