+++ shared/build.gradle.kts
@@ -40,6 +40,7 @@ kotlin {
                 implementation(projects.domain.lemmy.data)
                 implementation(projects.domain.lemmy.pagination)
                 implementation(projects.domain.lemmy.repository)
+                implementation(projects.domain.lemmy.usecase)
 
                 implementation(projects.unit.about)
                 implementation(projects.unit.accountsettings)
