+++ unit/explore/build.gradle.kts
@@ -31,6 +31,7 @@ kotlin {
                 implementation(projects.domain.lemmy.data)
                 implementation(projects.domain.lemmy.pagination)
                 implementation(projects.domain.lemmy.repository)
+                implementation(projects.domain.lemmy.usecase)
 
                 implementation(projects.unit.createcomment)
                 implementation(projects.unit.moderatewithreason)
