+++ unit/instanceinfo/build.gradle.kts
@@ -32,6 +32,7 @@ kotlin {
                 implementation(projects.domain.lemmy.data)
                 implementation(projects.domain.lemmy.pagination)
                 implementation(projects.domain.lemmy.repository)
+                implementation(projects.domain.lemmy.usecase)
             }
         }
     }
